refactor: 초기 폴더 구조 변경
