+++ src/main/java/dazaram/eureka/BaseTimeEntity.java
