+++ src/main/java/dazaram/eureka/ingredient/domain/Ingredient.java
