+++ src/main/java/dazaram/eureka/ingredient/domain/IngredientCategory.java
