+++ src/main/java/dazaram/eureka/recipe/domain/AiRecipe.java
