+++ src/main/java/dazaram/eureka/recipe/domain/Recipe.java
