+++ src/main/java/dazaram/eureka/recipe/domain/RecipeCategory.java
