+++ src/main/java/dazaram/eureka/recipe/domain/RecipePlatform.java
