+++ src/main/java/dazaram/eureka/recipe/domain/RecipeSequence.java
