+++ src/main/java/dazaram/eureka/recipe/repository/AiRecipeRepository.java
