+++ src/main/java/dazaram/eureka/user/domain/Taste.java
