+++ src/main/java/dazaram/eureka/user/domain/User.java
@@ -9,22 +9,28 @@ import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.OneToMany;
 
 import dazaram.eureka.BaseTimeEntity;
+import dazaram.eureka.ingredient.domain.CustomIngredient;
+import dazaram.eureka.ingredient.domain.UserIngredient;
 import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
+@Builder
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class User extends BaseTimeEntity {
 
 	@Id
-	@GeneratedValue
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "user_id")
 	private Long id;
 
@@ -34,24 +40,19 @@ public class User extends BaseTimeEntity {
 
 	private String profileImage;
 
-	@Enumerated(EnumType.STRING)
 	private Gender gender;
 
 	@OneToMany(mappedBy = "user", cascade = CascadeType.ALL)
-	private List<Oauth> oauths = new ArrayList<>();
+	private List<Oauth> oauths;
 
 	@OneToMany(mappedBy = "user", cascade = CascadeType.ALL)
-	private List<UserTaste> userTastes = new ArrayList<>();
+	private List<UserTaste> userTastes;
 
-	@Builder
-	public User(Long id, String name, String phoneNumber, boolean pushAlarmAllow, String profileImage, Gender gender) {
-		this.id = id;
-		this.name = name;
-		this.phoneNumber = phoneNumber;
-		this.pushAlarmAllow = pushAlarmAllow;
-		this.profileImage = profileImage;
-		this.gender = gender;
-	}
+	@OneToMany(mappedBy = "user", cascade = CascadeType.ALL)
+	private List<UserIngredient> userIngredients;
+
+	@OneToMany(mappedBy = "user", cascade = CascadeType.ALL)
+	private List<CustomIngredient> customIngredients;
 
 	public void addUserTaste(UserTaste userTaste) {
 		userTastes.add(userTaste);
@@ -59,5 +60,12 @@ public class User extends BaseTimeEntity {
 
 	public void addOauth(Oauth oauth) {
 		oauths.add(oauth);
+	}
+	public void addUserIngredient(UserIngredient userIngredient) {
+		userIngredients.add(userIngredient);
+	}
+
+	public void addCustomIngredient(CustomIngredient customIngredient) {
+		customIngredients.add(customIngredient);
 	}
 }
