+++ src/main/java/dazaram/eureka/user/domain/UserTaste.java
