+++ src/main/java/dazaram/eureka/user/repository/TasteRepository.java
