+++ src/main/java/dazaram/eureka/user/repository/UserTasteRepository.java
