+++ src/test/java/dazaram/eureka/EurekaApplicationTests.java
