+++ src/test/java/dazaram/eureka/ingredient/domain/IngredientCategoryTest.java
