+++ src/test/java/dazaram/eureka/recipe/domain/RecipeDomainTest.java
