+++ src/test/java/dazaram/eureka/user/repository/OauthRepositoryTest.java
