+++ src/test/java/dazaram/eureka/user/repository/TasteRepositoryTest.java
