+++ src/test/java/dazaram/eureka/user/repository/UserTasteRepositoryTest.java
